Notify user when a default config file is created
From time to time it'd create a default config even when I'd specify a
config path because the config path would not exist because of a type.
In these cases any of the tools using config would silently created a
config with default values in them, which is not what I expected.  Now
config tells you when it is creating a config file and also shows the
content of the file.

Also the print function to be used can be injected while instantiating
config.