Add goals
Goals is a separate table where you set your goals with a starting date,
end date, work-type and the number of hours you want to achieve before
the end date. It creates a Goals table to keep track of goals.
Everytime it shows you goals, it checks how many hours you have put in
since the starting of the goal and how much you still have to do and
what will be the average per day rate to achieve the goal.