Fix help option with argparse
argparse does not show the arguments and options that are added in main
functions when invoked with --help. The reason for this is that it exits
after parse_known_args. To fix this create a new ArgumentParser in main
function and make the earlier instance of the ArgumentParser as its
parent.